+++ src/app/(main)/market/_layout/Mobile/index.tsx
@@ -1,17 +1,25 @@
-import { PropsWithChildren } from 'react';
-
 import MobileContentLayout from '@/components/server/MobileNavLayout';
 
-import AgentSearchBar from '../../features/AgentSearchBar';
+import { LayoutProps } from '../type';
+import DetailModal from './DetailModal';
 import Header from './Header';
 
-const MobileLayout = ({ children }: PropsWithChildren) => {
+const Layout = ({ children, detail }: LayoutProps) => {
   return (
-    <MobileContentLayout gap={16} header={<Header />} style={{ paddingInline: 16 }} withNav>
-      <AgentSearchBar mobile />
-      {children}
-    </MobileContentLayout>
+    <>
+      <MobileContentLayout
+        gap={16}
+        header={<Header />}
+        style={{ paddingInline: 16, paddingTop: 8 }}
+        withNav
+      >
+        {children}
+      </MobileContentLayout>
+      <DetailModal>{detail}</DetailModal>
+    </>
   );
 };
 
-export default MobileLayout;
+Layout.displayName = 'MobileMarketLayout';
+
+export default Layout;