+++ src/locales/default/common.ts
@@ -28,8 +28,9 @@ export default {
   },
   feedback: '反馈与建议',
   follow: '在 {{name}} 上关注我们',
-  historyRange: '历史范围',
+  fullscreen: '全屏模式',
 
+  historyRange: '历史范围',
   import: '导入配置',
   importModal: {
     finish: {
@@ -133,6 +134,7 @@ export default {
   tab: {
     chat: '会话',
     market: '发现',
+    me: '我',
     setting: '设置',
   },
   telemetry: {
@@ -150,5 +152,19 @@ export default {
     action: '升级',
     hasNew: '有可用更新',
     newVersion: '有新版本可用：{{version}}',
+  },
+  userPanel: {
+    billing: '账单管理',
+    defaultNickname: '社区版用户',
+    discord: '社区支持',
+    docs: '使用文档',
+    email: '邮件支持',
+    feedback: '反馈与建议',
+    help: '帮助中心',
+    moveGuide: '设置按钮搬到这里啦',
+    plans: '订阅方案',
+    profile: '账户管理',
+    setting: '应用设置',
+    usages: '用量统计',
   },
 };